% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createIgnoreWords.R
\name{createIgnoreWords}
\alias{createIgnoreWords}
\title{List Ignored Words}
\usage{
createIgnoreWords(wordsList = c(), addToExisting = TRUE)
}
\arguments{
\item{wordsList}{List of words to be ignored when grouping user stories
semantically by similarities of word.}

\item{addToExisting}{If this param is TRUE, passed words will be added to a
predefined set of words. Otherwise, only those in the previous argument
will be ignored.}
}
\value{
Returns an array of words that will be ignored when processing the
    semantic groups.
}
\description{
List Ignored Words
}
\examples{
# Generating default words only
createIgnoreWords()

# Adding words
createIgnoreWords(c("given", "said"))

# Replacing words
createIgnoreWords(c("given", "said"), FALSE)

}
\seealso{
Other Simplified Process: 
\code{\link{analyseStories}()},
\code{\link{reportStories}()}
}
\concept{Simplified Process}
