% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_stories.R
\name{example_stories}
\alias{example_stories}
\title{Path to Example Data}
\usage{
example_stories()
}
\value{
Local path to example text file, containing user stories for Operational
    Research mathmatical models.
}
\description{
Path to Example Data
}
\examples{
example_stories()
}
\concept{Discogs data and functions}
