% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oRus.R
\docType{package}
\name{oRus}
\alias{oRus}
\title{\code{oRus} package}
\description{
Analysis of User Stories for Operational Research
}
\details{
See the README on \href{https://github.com/melvidoni/oRus}{GitHub}
}
