% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportStories.R
\name{reportStories}
\alias{reportStories}
\title{Reporting Stories}
\usage{
reportStories(
  stories,
  sheetFilePath = NULL,
  reportFilePath = NULL,
  outputType = "html_document"
)
}
\arguments{
\item{stories}{The stories produced by the analysis function.}

\item{sheetFilePath}{The path and filename of the Excel sheet that will be
stored; must include the `*.xlsx` extension. If no value is passed, the
file will not be written.}

\item{reportFilePath}{The path where the extensive report will be stored. It
must include the correct extension (according to the type selected in the
following argument). If no value is passed, the report won't be generated.}

\item{outputType}{The type of document to be generated (from an RMarkdown). By default
it is a PDF file. Options are: `html_output` or `pdf_output`.}
}
\description{
This function allows you to write the reports for the user stories
    if you didn't write them before on the analysis function. The key input
    is the output of `oRus::analyseStories()`.#'
}
\seealso{
Other Simplified Process: 
\code{\link{analyseStories}()},
\code{\link{createIgnoreWords}()}
}
\concept{Simplified Process}
